mix corne layout with myrioku
this is the combination of things I liked in myrioku layout, but with
the 42 keys. I feel like 34 layout is too extreme for me (at least for
now).
however, I like the idea of separating layers more, and having most
layers enabled cross-hand.
so I took all ideas I like in miryoku, and applied them to the previous
cornified layout I had.