change thumb cluster declaration in layers
turns out  that &trans delegates key press handling to the next active
layer. That means that the order of their declaration in the keymap
doesn't affect the execution, and active layer stack is decided in
runtime.
Because of that, I changed thunb cluster definition in all layers to
avoid triggering different layer while holding any thumb key.