rever layer ordering
turns out layers stop behaving as I expect once sys layer is put beyond
sym and hub layers. So I added a &to key to the sys layer to provide
the same result - be able to quit sys layer through hub - with the same
key combo required to do that through sym layer.