change layer taps into combo words
I stopped using my split keyboard altogether because it was causing me
pain in my right shoulder blade. I suspect that is because I
have to hold some buttons to receive a layer I need.
I want to try to circumvent that using urob's sticky layers - num word
and newly created nav word.
NUM layer in turn is also changed to be on a home row, to reduce the
amount of travel required to use it
NAV layer is left the same, but can now only be triggered via the new
nav word behaviour through combo
The CANCEL combo is also introduced to be able to get out of sticky
layers In case I need to type a button that is in the continue list for
the current sticky layer